server/requirements.txt
flask
flask-restful
flask-jwt-extended
flask-sqlalchemy
flask-migrate
sqlalchemy-serializer
twilio
psycopg2-binary
openai
python-dotenv
spacy
en-core-web-sm
alembic==1.13.3
aniso8601==9.0.1
anyio==3.7.1
async-property==0.2.2
attrs==24.2.0
Authlib==1.3.2
bcrypt==4.2.0
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
cached-property==1.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
confection==0.1.5
cors==1.0.1
cryptography==43.0.3
cymem==2.0.10
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.9
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==30.8.1
filelock==3.16.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
flask-oidc==2.2.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
future==1.0.0
fuzzywuzzy==0.18.0
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.12.0
httpcore==0.13.7

idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
jwt==1.3.1
langcodes==3.4.1
language_data==1.3.0
Levenshtein==0.25.1
limits==3.13.0
Mako==1.3.5
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mistune==3.0.2
murmurhash==1.0.11
mysqlclient==2.2.5
numpy==1.21.0
ordered-set==4.1.0
packaging==24.1
passlib==1.7.4
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.4.0
pipenv==2024.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.20
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-Levenshtein==0.25.1
pytz==2024.2
PyYAML==5.4.1
rapidfuzz==3.9.7
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
SQLAlchemy-serializer==1.4.12
srsly==2.4.8
svgwrite==1.4.3
thinc==8.1.12
tldextract==5.1.2
tomli==2.0.2
tqdm==4.67.1
Tree==0.2.4
typer==0.7.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
wasabi==1.1.3
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.1
# python -m venv supabase-env
# pip install supabase


